+++ apps/website/src/app/(content)/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default async function Page() {
+  return (
+    <div>
+      <h1>Home</h1>
+    </div>
+  );
+}