+++ apps/website/src/components/content/content-filter.tsx
@@ -1,0 +1,21 @@
+'use client';
+
+<<<<<<< Updated upstream
+import React, { FunctionComponent } from 'react';
+=======
+import React from 'react';
+>>>>>>> Stashed changes
+
+type ContentFilterProps = {
+  // TODO: Add props here.
+};
+
+<<<<<<< Updated upstream
+export const ContentFilter: FunctionComponent<ContentFilterProps> = ({}: ContentFilterProps): React.JSX.Element => {
+  return <div></div>;
+};
+=======
+export function ContentFilter({}: ContentFilterProps): React.JSX.Element {
+  return <div></div>;
+}
+>>>>>>> Stashed changes