+++ apps/website/src/lib/auth/sign-in-with-email-and-password.action.ts
@@ -1,0 +1,28 @@
+'use server';
+
+<<<<<<< Updated upstream
+import { SignInFormState } from '@/lib/auth/sign-in-form-state';
+
+export async function signInWithEmailAndPassword(previousState: SignInFormState, formData: FormData): Promise<SignInFormState> {
+  // TODO: Implement sign-up with email and password.
+=======
+import { httpClientFactory } from '@/lib/core/http';
+import { SignInFormState } from '@/lib/auth/sign-in-form-state';
+import { SignInWithEmailAndPasswordGateway } from '@/lib/auth/sign-in-with-email-and-password.gateway';
+
+export type SignInWithEmailAndPasswordAction = {
+  signInWithEmailAndPassword: (email: Readonly<string>, password: Readonly<string>) => Promise<any>;
+};
+
+export async function signInWithEmailAndPassword(previousState: SignInFormState, formData: FormData): Promise<SignInFormState> {
+  // TODO: Implement sign-up with email and password.
+  const gateway = new SignInWithEmailAndPasswordGateway(httpClientFactory());
+
+  const email = formData.get('email') as string;
+  const password = formData.get('password') as string;
+
+  const response = await gateway.signInWithEmailAndPassword(email, password);
+
+>>>>>>> Stashed changes
+  return Promise.resolve({});
+}